changing datatypes, ranking scheme

+++ FileSearchEngine/src/filesearchengine/CMDQueryProcess.java
@@ -21,7 +21,7 @@ public class CMDQueryProcess {
         super();
     }
     
-    private static void displayResult(Map<Integer, Float> docScoreMap, Map<Integer, DocInfo> docIdFileMap){
+    private static void displayResult(Map<Integer, Double> docScoreMap, Map<Integer, DocInfo> docIdFileMap){
         System.out.println("*****************Displaying the Final result*****************");
         if(docScoreMap != null && !docScoreMap.isEmpty()){
             for(Integer docId : docScoreMap.keySet()){
@@ -49,7 +49,7 @@ public class CMDQueryProcess {
         MainQueryProcess mainProc = new MainQueryProcess(corpusInfo);
         
         //Query with search terms
-        Map<Integer, Float> docScoreMap = mainProc.triggerQuery(args[1]);
+        Map<Integer, Double> docScoreMap = mainProc.triggerQuery(args[1]);
         //Display result in order of relevance
         CMDQueryProcess.displayResult(CommonUtils.sortByValue(docScoreMap, 2/*fetch top results*/), corpusInfo.getDocIdInfoMap());
     }
